\docType{package}
\name{DNAprofiles-package}
\alias{DNAprofiles-package}
\title{DNA profile analysis}
\description{
  Methods for studying forensic DNA profiles, in particular
  familial databases searches.
}
\details{
  More detail (?) ...
}
\author{
  \email{m.v.kruijver@vu.nl}
}
\keyword{package}

