% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAprofiles-package.R
\docType{data}
\name{freqsNLngm}
\alias{freqsNLngm}
\title{Allelic frequencies at 15 NGM loci}
\format{
A list containing a vector (loci) and a sub-list (freqs).
}
\source{
Netherlands Forensic Institute
}
\usage{
freqsNLngm
}
\description{
Allelic frequencies from a reference population (consisting of 2,085 Dutch males) used by the Netherlands Forensic Institute.
}
\seealso{
\code{\link{freqsNLsgmplus}},\code{\link{freqsUScaucs}}
}
