\docType{data}
\name{freqsNLsgmplus}
\alias{freqsNLsgmplus}
\title{Allelic frequencies at 10 SGMplus loci}
\format{A list containing a vector (loci) and a sub-list (freqs).}
\source{
Netherlands Forensic Institute
}
\usage{
freqsNLngm
}
\description{
Allelic frequencies from a reference population (consisting
of 2,085 Dutch males) used by the Netherlands Forensic
Institute.
}
\seealso{
\code{\link{freqsNLngm}},\code{\link{freqsUScaucs}}
}

