\docType{data}
\name{freqsUScaucs}
\alias{freqsUScaucs}
\title{Allelic frequencies at 13 CODIS loci}
\format{A list containing a vector (loci) and a sub-list (freqs).}
\source{
Copied from the relSim package
}
\usage{
freqsNLngm
}
\description{
Allelic frequencies from a US reference population. This
dataset is copied from the \code{relSim} package, raw data
is available from
\url{http://www.fbi.gov/about-us/lab/forensic-science-communications/fsc/july1999/dnaloci.txt}
}
\seealso{
\code{\link{freqsNLsgmplus}},\code{\link{freqsNLngm}}
}

