% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get.freqs}
\alias{get.freqs}
\title{Retrieve the allele frequencies of a profiles object}
\usage{
get.freqs(x)
}
\arguments{
\item{x}{profiles object}
}
\description{
A \code{\link{profiles}} object is an integer matrix with two columns per marker. This function extracts the allele frequencies from the attributes of the profiles object.
}
\examples{
data(freqsNLsgmplus)
           x<- sample.profiles(1,freqsNLsgmplus)
           stopifnot(identical(get.freqs(x),freqsNLsgmplus))
}
