% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{get.markers}
\alias{get.markers}
\title{Retrieve the markers of a profiles object}
\usage{
get.markers(x)
}
\arguments{
\item{x}{profiles object}
}
\description{
A \code{\link{profiles}} object is an integer matrix with two columns per marker. This function extracts the marker names from the column names of the profiles object.
}
\examples{
data(freqsNLsgmplus)
           x<- sample.profiles(1,freqsNLsgmplus)
           stopifnot(identical(get.markers(x),names(freqsNLsgmplus)))
}
