% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize.freqs}
\alias{normalize.freqs}
\title{Normalizes allele frequencies such that their sum is 1}
\usage{
normalize.freqs(freqs)
}
\arguments{
\item{freqs}{list of per locus allele frequencies}
}
\value{
list
}
\description{
Normalizes allele frequencies such that their sum is 1
}
\examples{
data(freqsNLsgmplus)
fr0 <- normalize.freqs(freqsNLsgmplus)
stopifnot(all(sapply(fr0,sum)==1))
}

