% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{profiles}
\alias{profiles}
\title{profiles object}
\usage{
profiles(x, labels)
}
\arguments{
\item{x}{list of character vectos with genotypes}

\item{labels}{allele levels per locus}
}
\description{
Profiles are stored in a profiles object, which is merely an integer matrix. The profiles function creates such a matrix from the genotypes of profiles stored in a list. Alleles are encoded as integers following the order in the labels argument. See the example for details.
}
\examples{
data(freqsNLngm)
x <- profiles(list(D1S1656="12/15",D2S441="11/14",D2S1338="19/19",D3S1358="16/18",
                   FGA="22/24",D8S1179="11/12",D10S1248="13/14",TH01="7/9",VWA="16/17",
                   D12S391="17/20",D16S539="12/13",D18S51="13/17",
                   D19S433="13/14",D21S11="27/31.2",D22S1045="15/16"),
             labels = get.labels(freqs = freqsNLngm))
x
}
