% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{freqsNLngm}
\alias{freqsNLngm}
\title{Allelic frequencies at 15 NGM loci}
\format{A list containing a vector (loci) and a sub-list (freqs).}
\source{
Netherlands Forensic Institute
}
\usage{
freqsNLngm
}
\description{
Allelic frequencies from a reference population (consisting of 2,085 Dutch males) used by the Netherlands Forensic Institute.
}
\seealso{
\code{\link{freqsNLsgmplus}},\code{\link{freqsUScaucs}}
}

