\docType{data}
\name{ibdprobs}
\alias{ibdprobs}
\title{IBD probabilites for pairwise relationships (FS, PO, etc.)}
\format{A list containing vectors with IBD-probabilities.}
\description{
  IBD probabilities for the following pairwise
  relationships: \itemize{ \item "FS", full siblings \item
  "PO", parent/offspring \item "HS", half siblings \item
  "AV", uncle/nephew \item "FC", first cousins \item "SC",
  second cousins }
}
\keyword{datasets}

