% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normalize.freqs}
\alias{normalize.freqs}
\title{Normalizes allele frequencies such that their sum is 1}
\usage{
normalize.freqs(freqs)
}
\arguments{
\item{freqs}{list of per locus allele frequencies}
}
\value{
list
}
\description{
Normalizes allele frequencies such that their sum is 1
}
\examples{

data(freqsNLsgmplus)
fr0 <- normalize.freqs(freqsNLsgmplus)
stopifnot(all.equal(sapply(fr0,sum),setNames(rep(1,length(fr0)),names(fr0))))
}
